dts: scobc_a1_dev: Add LED/GPIO definitions
Add LED and GPIO definitions to shield dts overlay.
The SC OBC Module A1 is equipped with AMD (formerly Xilinx)'s AXI
GPIO IP core, which connects to 10 GPIOs. The relationship between the
GPIOs and the User IO pins is as follows, with GPIO7 connected to the
LED on the Development Board.

  - UIO2_06: GPIO6
  - UIO2_07: GPIO7
  - UIO2_08: GPIO8
  - UIO2_09: GPIO9
  - UIO2_10: GPIO10
  - UIO2_11: GPIO11
  - UIO2_12: GPIO12
  - UIO2_13: GPIO13
  - UIO2_14: GPIO14
  - UIO2_15: GPIO15

The alias for `led0` has been added to allow the Zephyr's
`samples/basic/blinky` to be tested immediately.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>