dts: scobc_a1_dev: Add I2C definitions
Add I2C definitions to shield DTS overlay.
These I2C is connected to the following pins on the SC-OBC Module A1
Development Board.

  - i2c0:
    - I2C_EXT_SCL: I2C_SCK
    - I2C_EXT_SDA: I2C_SDA

  - i2c1
    - UIO1_12:_I2C_SCK
    - UIO1_13: I2C_SDA

  - i2c2
    - UIO1_14:_I2C_SCK
    - UIO1_15: I2C_SDA

`i2c0` is connected to the temperature sensor (TMP175) on the SC-OBC
Module A1 Development Board.

The IP core that controls these I2C is developed by Space Cubics, and
the driver will be added in a separate commit.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>