west.yaml: Temporarily change the Zephyr repository
The current Zephyr upstream board definition for the SC-OBC Module
A1 targets the EM (Engineering Model), and defines a reduced number
of IRQs. As a result, the build test fails.

Since the PR to update the board definition for the FM (Flight Model)
is still under review, we will temporarily point `west.yaml` to the
review branch.
Once the changes are merged upstream, `west.yaml` should be reverted
to its original state.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>