bootloader: Add simple bootloader
This commit adds a simple bootloader.
This simple bootloader directly accesses the registers of the FPGA
SPI Core without using the SPI driver in Zephyr.

And the bootloader determines whether the FPGA is booting from the
update or golden image. Based on the result, the bootloader decide
which Configuration Memory address to load.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>